% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{barsFun}
\alias{barsFun}
\title{Create labelled barplot}
\usage{
barsFun(bars, main, xlab, barNames)
}
\arguments{
\item{bars}{vector of values to plot}

\item{main}{name of barplot for main title}

\item{xlab}{label for x-axis}

\item{barNames}{labels for each bar}
}
\description{
This function allows you to Create labelled barplot.
}
\examples{
barsFun()

}
\keyword{cats}
