% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{capXtreme}
\alias{capXtreme}
\title{cap extreme values in dataset}
\usage{
capXtreme(formulas, m, dataset, orderCol)
}
\arguments{
\item{formulas}{formulas dataframe.}

\item{m}{model iteration.}

\item{dataset}{dataset for model input.}

\item{orderCol}{name of column used to re-order the dataset (as.character) eg. orderCol = "X".}
}
\description{
This function allows you to cap extreme values in dataset.
}
\examples{
capXtreme()

}
\keyword{cats}
