% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{checks}
\alias{checks}
\title{Wrapper function to check for warnings and errors before committing}
\usage{
checks(wd = "C:\\\\Users\\\\Rose\\\\MScResPACK\\\\MScResPACK")
}
\arguments{
\item{wd}{path to find files that need checking.}
}
\description{
This function allows you to check for warnings and errors before committing changes. Depends on devtools.
}
\examples{

checks()

}
\keyword{cats}
