% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourMapper.R
\name{colourMapper}
\alias{colourMapper}
\title{Plot up to 2 rasters}
\usage{
colourMapper(topmapDF, colours)
}
\arguments{
\item{topmapDF}{Dataframe object containing "Val", "Lon" and "Lat".}

\item{colours}{Dataframe object containing "colLeg" (a vector of colours) and "t" (a vector of regular intervals between minimum and maximum of "topmapDF$Val")}
}
\description{
This function allows you to create string of colours for a numeric vector, based on user-specified bins.
}
\examples{
colourMapper()
}
\keyword{cats}
