% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{covStrucData}
\alias{covStrucData}
\title{Set up for dataset for covariance structures}
\usage{
covStrucData(dataset, strucType)
}
\arguments{
\item{dataset}{dataset used for spatial covariance xy.}

\item{strucType}{either "temporal" or "spatial".}
}
\description{
This function allows you to set up dataset for covariance structures.
}
\examples{
covStrucData(dataset, strucType="temporal")

}
\keyword{cats}
