% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{devROxWrap}
\alias{devROxWrap}
\title{creating or updating .Rd files for functions}
\usage{
devROxWrap(wd = "M:/My Documents/MScResPACK/MScResPACK")
}
\arguments{
\item{wd}{working directory path from which the file to document exists and .Rd file will be created. Default is "M:/My Documents/MScResPACK/MScResPACK"}
}
\description{
This function allows you to create or update .Rd files for functions.
}
\examples{

devROxWrap()

}
\keyword{cats}
