% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{effectExtensFun}
\alias{effectExtensFun}
\title{create vector of effects and/or extensions for model}
\usage{
effectExtensFun(formulas, m)
}
\arguments{
\item{formulas}{formulas dataframe.}

\item{m}{model iteration.}
}
\description{
This function allows you to create vector of effects and/or extensions for model.
}
\examples{
effectExtensFun()

}
\keyword{cats}
