% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{formFun}
\alias{formFun}
\title{write formula for model}
\usage{
formFun(formulaType = c("cond"), response = NULL, predictors, effExt)
}
\arguments{
\item{formulaType}{defaults to "cond".}

\item{response}{name of response variable (as character), defaults to NULL.}

\item{predictors}{name of predictor variables (as character)}

\item{effExt}{offset or autocorrelation structure.}
}
\description{
This function allows you to write formula for model.
}
\examples{
formFun()

}
\keyword{cats}
