% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kFoldCV.R
\name{kFoldCV}
\alias{kFoldCV}
\title{K-fold cross-validation}
\usage{
kFoldCV(sppColRef, dataset, NameFull, k)
}
\arguments{
\item{sppColRef}{Name of species column. Also used as abbreviation of names for references.}

\item{dataset}{Dataframe object to be used in k-fold cross validation models.}

\item{NameFull}{Full name of the species}

\item{k}{Number the dataset should be split into and model iterations}
}
\description{
This function allows you to carry out K-fold cross-validation, specifying k.
}
\examples{
kFoldCV()
}
\keyword{cats}
