% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map function.R
\name{mapFun}
\alias{mapFun}
\title{Plot up to 2 rasters}
\usage{
mapFun(baseRefsDf = NULL, legTOP = NULL, mapsVis = "both",
  basemapOutline = NULL, basemapDF = NULL, topmapDF = NULL,
  wdExtension = NULL, mapName = NULL, countOnly = FALSE,
  bubble = TRUE)
}
\arguments{
\item{baseRefsDf}{(dataframe) dataframe of basemap refs; names to call in (df[,1] = "fileNm") and legend label (df[,2] = "leglab")}

\item{legTOP}{(character) label of legend for top map}

\item{mapsVis}{("base","top","topWbaseCont","both") which map should be plotted; default = "both", topWbaseCont means plot top map with contours from base raster}

\item{basemapOutline}{(character or NULL) name of basemap outline to call in (optional)}

\item{basemapDF}{raster object or null for basemap}

\item{topmapDF}{(character) name of dataframe for top map, col names must be ("Val", "Lon", "Lat")}

\item{wdExtension}{(character) location within the working directory to save the map (must end in "/")}

\item{mapName}{(character) file name}

\item{countOnly}{(logical, TRUE/FALSE) for maps where no zeros are desired, if TRUE, mapFun subsets values > 0 (for raster)}

\item{bubble}{(logical, TRUE/FALSE) default = TRUE, should top map be bubble or normal raster}
}
\description{
This function allows you to plot with max=2 rasters: basemap and top map, with options for contours.
}
\examples{
mapFun()
}
\keyword{cats}
