% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{modelFun}
\alias{modelFun}
\title{Create model}
\usage{
modelFun(fams, fam, linkTypes, link, dataset, fcond, fzeroi)
}
\arguments{
\item{fams}{}

\item{fam}{}

\item{linkTypes}{}

\item{link}{}

\item{dataset}{}

\item{fcond}{}

\item{fzeroi}{}
}
\description{
This function allows you to write a glmmTMB model.
}
\examples{
modelFun()

}
\keyword{cats}
