% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{parametersPlotFun}
\alias{parametersPlotFun}
\title{Plot Fitted vs explanatory variables: conditional/zero inflated}
\usage{
parametersPlotFun(y, data, parameters, paramAbbrv, paramFullNam, fullPath,
  modelRefNo, modelRefFull, yName, paramTypeNam)
}
\arguments{
\item{y}{vector for y axis (fitted values)}

\item{data}{dataframe containing explanatory variables}

\item{parameters}{parameters}

\item{paramAbbrv}{parameters abbreviation}

\item{paramFullNam}{full name of parameters}

\item{fullPath}{Full file pathway to save .png}

\item{modelRefNo}{model reference number}

\item{modelRefFull}{full model reference}

\item{yName}{name for y axis}

\item{paramTypeNam}{name of parameter type}
}
\description{
This function allows you to Plot Fitted vs explanatory variables: conditional/zero inflated.
}
\examples{
parametersPlotFun()

}
\keyword{cats}
