% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparing abundance dataset.R
\name{prepDF}
\alias{prepDF}
\title{Add dummy variable to dataset and convert the dummy, month and season variables to factors}
\usage{
prepDF(dataset)
}
\arguments{
\item{dataset}{dataframe used to create dummy variable to dataset and convert the dummy, month and season variables to factors}
}
\description{
This function allows you add dummy variable to dataset and convert the dummy, month and season variables to factors.
}
\examples{
prepDF()


}
\keyword{cats}
