% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastFun.R
\name{rastFun}
\alias{rastFun}
\title{wrapper function for sp and raster}
\usage{
rastFun(x, y, z, overZero = FALSE)
}
\arguments{
\item{x}{vector of Longitude}

\item{y}{vector of Latitude}

\item{z}{vector of values to be rasterised}

\item{overZero}{(logical, TRUE/FALSE) should rastFun subset data over zero, default == FALSE}
}
\description{
This function allows you to create a raster object.
}
\examples{
rastFun()


}
\keyword{cats}
