% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparing abundance dataset.R
\name{seasonsIndFun}
\alias{seasonsIndFun}
\title{Create index of seasons in abundance dataset}
\usage{
seasonsIndFun(dataset)
}
\arguments{
\item{dataset}{dataframe used to create seasons index}
}
\description{
This function allows you create index of seasons in abundance dataset.
}
\examples{
seasonsIndFun()

}
\keyword{cats}
