% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeSimEq.R
\name{slopeSimEq}
\alias{slopeSimEq}
\title{scale a vector 0 to 1}
\usage{
slopeSimEq(x1 = 0.1, x2 = 0.9, c, y, m, scaleMin, scaleMax, file)
}
\arguments{
\item{x1}{value 1 of explanatory variable (for dy/dx)}

\item{x2}{value 2 of explanatory variable (for dy/dx)}

\item{c}{Intercept}

\item{y}{response variable}

\item{m}{Vector of unscaled slope values}

\item{scaleMin}{minimum value to scale to}

\item{scaleMax}{maximum value to scale to}

\item{file}{file pathway and file name to save plot to}
}
\description{
This function allows you to scale a vector 0 to 1.
}
\examples{
slopeSimEq()

########
# data #
########

condSummary1 = read.csv("C:/Users/Rose/Documents/Rose/Abundance models- whole study area/Species/Cetaceans/Bottlenose Dolphin/glmmTMB/Model objects/Coefficients/BTNDtmb5.1.1.1 coefficients table _ Conditional.csv",stringsAsFactors = FALSE)
condSummary2 = read.csv("C:/Users/Rose/Documents/Rose/Abundance models- whole study area/Species/Cetaceans/Bottlenose Dolphin/glmmTMB/Model objects/Coefficients/BTNDtmb5.2.1.1 coefficients table _ Conditional.csv",stringsAsFactors = FALSE)

sppColRef=formulas$dataset_name[m]

dataset = capXtreme(formulas,m,dataset,"X")

##############
# Parameters #
##############

c1 = condSummary1$Estimate[1]
c2 = condSummary2$Estimate[1]

y = dataset$BTND1

estLen = length(condSummary1$Estimate)
m1 = c(condSummary1$Estimate[2:estLen],0)

estLen = length(condSummary2$Estimate)
m2 = c(condSummary2$Estimate[2:estLen],0)
rm(estLen)

scaleMin = -1
scaleMax = 1


poiss = slopeSimEq(x1=0.2,x2=0.7,c=c1,y=y,m=m1,scaleMin = scaleMin,scaleMax = scaleMax)

negbin1 = slopeSimEq(x1=0.2,x2=0.7,c=c2,y=y,m=m2,scaleMin = scaleMin,scaleMax = scaleMax)

poiss$fam = "poiss"
poiss$cols = "deeppink"

negbin1$fam = "negbin1"
negbin1$cols = "blue"


bothModels = rbind(poiss,negbin1)

##############
# plot #
############

op = par(mfrow = c(3,2))

plot(bothModels$slopesScaled,col = bothModels$cols,pch = 20,cex = 2,main = "slope scaled")
abline(h=bothModels$slopesScaled[6],col = "deeppink")
abline(h=bothModels$slopesScaled[12],col = "blue",lty = "dashed")

boxplot(bothModels$slopesScaled~bothModels$fam,col=c("blue","deeppink"),main = "slope scaled")


plot(bothModels$m,col = bothModels$cols,pch = 20,cex = 2,main = "original slope value (m)")
abline(h=bothModels$m[6],col = "deeppink")
abline(h=bothModels$m[12],col = "blue",lty = "dashed")


boxplot(bothModels$m~bothModels$fam,col=c("blue","deeppink"),main = "original slope value (m)")


plot(bothModels$distFrm0,col = bothModels$cols,pch = 20,cex = 2,main = "distFrm0")
abline(h=bothModels$distFrm0[6],col = "deeppink")
abline(h=bothModels$distFrm0[12],col = "blue",lty = "dashed")


boxplot(bothModels$distFrm0~bothModels$fam,col=c("blue","deeppink"),main = "distFrm0")


par(op)


plot(bothModels$distFrm0~bothModels$m,col=c("deeppink","blue"),main = "distFrm0 ~ original slope value (m)")
abline(h=0,col = "darkorange",lty = "dashed")
abline(v=0,col = "darkgreen",lty = "dashed")



}
\keyword{cats}
