% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{spACsetup}
\alias{spACsetup}
\title{Set up for spatial autocorrelation structure}
\usage{
spACsetup(dataset)
}
\arguments{
\item{dataset}{dataset used for spatial covariance xy.}
}
\description{
This function allows you to Set up for spatial autocorrelation structure, by creating pos.
}
\examples{
spACsetup()

}
\keyword{cats}
