% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{summaryTblFun}
\alias{summaryTblFun}
\title{create empty summary table}
\usage{
summaryTblFun(tableLength)
}
\arguments{
\item{tableLength}{number of rows.}
}
\description{
This function allows you to create empty summary table.
}
\examples{
summaryTblFun()

}
\keyword{cats}
