% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{sysTimeDiff}
\alias{sysTimeDiff}
\title{end Sys.time and return difference}
\usage{
sysTimeDiff(start)
}
\arguments{
\item{start}{object of Sys.time set before running command to be timed.}
}
\description{
This function allows you to end Sys.time and return the time difference.
}
\examples{

# create start time object
start = Sys.time()

# run command to be timed
print("hello")


sysTimeDiff(start)

}
\keyword{cats}
