% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{tableLenFun}
\alias{tableLenFun}
\title{find summary table length}
\usage{
tableLenFun(fams, effortVec)
}
\arguments{
\item{fams}{index for families.}

\item{effortVec}{index for effort.}
}
\description{
This function allows you to work out how long the summary table should be.
}
\examples{
tableLenFun()

}
\keyword{cats}
