% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropRW.R
\name{metropRW}
\alias{metropRW}
\title{Random walk metropolis sampling wrapper function}
\usage{
metropRW(model, filePath)
}
\arguments{
\item{model}{model object of glmmTMB class}

\item{filePath}{file pathway, must end with "/" or beginning of the file name}
}
\description{
This function allows you to carry out Random Walk Metroplis Sampling on a model, and creates two plots.
}
\details{
Returns list object containing run time, isEqual ()
m1,isEqual,timer
}
\examples{
filePath = paste(Path,"Plots/",ModelRefNo, sep = "")
metropRW(model,filePath)

}
\keyword{cats}
