% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSpecificFuns.R
\name{modelFun}
\alias{modelFun}
\title{Create model}
\usage{
modelFun(fams, fam, linkTypes, link, dataset, fcond, fzeroi)
}
\arguments{
\item{fams}{family type string}

\item{fam}{family iteration}

\item{linkTypes}{link type string}

\item{link}{link iteration}

\item{dataset}{dataset to be used in model}

\item{fcond}{formula for conditional model}

\item{fzeroi}{formula for logistic model}
}
\description{
This function allows you to write a glmmTMB model.
}
\examples{
modelFun()

}
\keyword{cats}
