% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasOutline function.R
\name{rasOutline}
\alias{rasOutline}
\title{A Cat Function}
\usage{
rasOutline(x, plot = TRUE, installRgeos = FALSE)
}
\arguments{
\item{plot}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
rasOutline()

}
\keyword{cats}
